requirements.txt
Package                       Version
----------------------------- --------
asgiref                       3.8.1
certifi                       2024.6.2
cffi                          1.16.0
charset-normalizer            3.3.2
cryptography                  42.0.7
defusedxml                    0.8.0rc2
Django                        5.0.6
django-debug-toolbar          4.4.2
django-templated-mail         1.1.1
djangorestframework           3.15.1
djangorestframework-simplejwt 5.3.1
djoser                        2.2.2
idna                          3.7
mysqlclient                   2.2.4
oauthlib                      3.2.2
pip                           22.0.2
pycparser                     2.22
PyJWT                         2.8.0
python-dotenv                 1.0.1
python3-openid                3.2.0
requests                      2.32.3
requests-oauthlib             2.0.0
setuptools                    59.6.0
social-auth-app-django        5.4.1
social-auth-core              4.5.4
sqlparse                      0.5.0
typing_extensions             4.12.1
urllib3                       2.2.1
wheel                         0.37.1
